requirements.txt
<<<<<<< HEAD
aiohttp==3.7.4
aiohttp-cors==0.7.0
aioredis==1.3.1
ansible==2.9.9
apache-libcloud==3.0.0
argcomplete==1.8.1
astroid==2.5
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
blessings==1.7
blinker==1.4
blis==0.7.4
boto3==1.4.8
botocore==1.8.50
bpytop==1.0.62
Brlapi==0.7.0
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.4.5.1
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
colorful==0.5.4
command-not-found==0.3
cryptography==3.0
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.5
dask==2.19.0
dbus-python==1.2.16
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.0.0/de_core_news_sm-3.0.0-py3-none-any.whl
decorator==4.4.2
defer==1.0.6
distributed==2.19.0
distro==1.5.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
ecdsa==0.15
filelock==3.0.12
Flask==1.1.2
fuzzysearch==0.7.3
google-api-core==1.26.0
google-auth==1.27.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
greenlet==0.4.15
grip==4.5.2
grpcio==1.35.0
HeapDict==1.0.1
hidpidaemon==18.4.6
hiredis==1.1.0
html5lib==1.1
httplib2==0.18.1
idna==2.10
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kernelstub==3.1.4
keyring==21.3.0
kitty-theme-changer @ git+git://github.com/fretboardfreak/kitty-theme-changer.git@4d38b85dfda44ddc9b0f6fda986d1c42b3e9db55
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lazy-object-proxy==1.5.2
lockfile==0.12.2
louis==3.14.0
lxml==4.5.2
macaroonbakery==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
modin @ git+https://github.com/modin-project/modin@057eb61dea345f2fce70e9990b775e466cf48fae
msgpack==1.0.2
multidict==5.1.0
murmurhash==1.0.5
netaddr==0.7.19
netifaces==0.10.4
ntlm-auth==1.4.0
numpy==1.18.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.12
opencensus-context==0.1.2
packaging==20.9
pandas==1.2.2
parso==0.8.1
path-and-address==2.0.1
pathy==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plumbum==1.7.0
pop-transition==1.1.2
preshed==3.0.5
progress==1.5
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.12.3
psutil==5.8.0
ptyprocess==0.7.0
py-spy==0.3.4
pyaes==1.6.1
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.9.7
pycups==2.0.1
pydantic==1.7.3
pydbus==0.6.0
Pygments==2.8.0
PyGObject==3.38.0
PyJWT==1.7.1
pykerberos==1.1.14
pylint==2.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pynvim==0.4.1
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.1.3+ubuntu1.3
python-dateutil==2.8.1
python-debian==0.1.37
python-dotenv==0.15.0
python-xlib==0.27
pytz==2020.1
pywinrm==0.3.0
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
ray==1.2.0
redis==3.5.3
repolib==1.5.2
repoman==1.4.0
requests==2.23.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-unixsocket==0.2.0
rpyc==4.1.5
rsa==4.7.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.1.13
scipy==1.6.1
scour==0.37
seaborn==0.11.1
SecretStorage==3.1.2
selinux==3.1
sessioninstaller==0.0.0
simplejson==3.17.0
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.3.0
soupsieve==2.0.1
spacy==3.0.3
spacy-legacy==3.0.1
srsly==2.4.0
systemd-python==234
tblib==1.7.0
Telethon==1.19.5
thinc==8.0.1
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.57.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.9
vboxapi==1.0
wadllib==1.3.4
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wordcloud==1.8.1
wrapt==1.12.1
xkit==0.0.0
xmltodict==0.12.0
yarl==1.6.3
youtube-dl==2020.9.14
zict==2.0.0
=======
bpytop==1.0.62
progress==1.5
psutil==5.8.0
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.15.0
rsa==4.7.1
ruamel.yaml==0.16.12
Telethon==1.19.5
>>>>>>> a48f43aeda54445fd1c57754ad6d3af9808b4c68
